acadpredict/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
pyjwt==2.9.0
joblib==1.4.2
python-multipart==0.0.9
pydantic==2.9.2
